Move crypto state out of device
