+++ wireguard-rs/src/wireguard/router/mod.rs
@@ -1,5 +1,6 @@
 mod anti_replay;
 mod constants;
+mod crypto_state;
 mod device;
 mod ip;
 mod messages;
